New hml

+++ src/controllers/estimates/deleteEstimateController.ts
@@ -16,6 +16,15 @@ export class DeleteEstimateController {
         const estimate = await prisma.estimate.findUnique({
             where: {
                 id: estimateId
+            },
+            select: {
+                status: true,
+                projectId: true,
+                project: {
+                    select: {
+                        status_project: true,
+                    }
+                }
             }
         })
 
@@ -25,10 +34,22 @@ export class DeleteEstimateController {
             })
         }
 
+        if (estimate.status === "approved") {
+            return res.status(400).json({
+                error: "It is not possible to delete an approved estimate."
+            })
+        }
+
         try {
             await prisma.estimate.delete({
                 where: {
                     id: estimateId
+                }
+            })
+
+            await prisma.project.delete({
+                where: {
+                    id: estimate.projectId
                 }
             })
